accessible/tests/mochitest/selectable/test_menu.xhtml

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<?xml-stylesheet href="../treeview.css" type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="XUL tree selectable tests">

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />

  <script type="application/javascript"
          src="../common.js" />
  <script type="application/javascript"
          src="../role.js" />
  <script type="application/javascript"
          src="../states.js" />
  <script type="application/javascript"
          src="../selectable.js" />

  <script type="application/javascript">
  <![CDATA[

    ////////////////////////////////////////////////////////////////////////////
    // Test

    //gA11yEventDumpID = "debug";

    var gQueue = null;

    function doTest()
    {
      //////////////////////////////////////////////////////////////////////////
      // menu

      var id = "menu";
      var menu = getAccessible("menu");
      var menuList = menu.firstChild;
      todo(isAccessible(menuList, [nsIAccessibleSelectable]),
           "No selectable accessible for list of menu '" + id + "'");

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  ]]>
  </script>

  <hbox flex="1" style="overflow: auto;">
    <body xmlns="http://www.w3.org/1999/xhtml">
      <a target="_blank"
        href="https://bugzilla.mozilla.org/show_bug.cgi?id=590176"
        title="add pseudo SelectAccessible interface">
       Mozilla Bug 590176
      </a><br/>
      <p id="display"></p>
      <div id="content" style="display: none">
      </div>
      <pre id="test">
      </pre>
    </body>

    <vbox flex="1">
      <menu label="menu" id="menu">
        <menupopup>
          <menuitem label="item1" id="m_item1"/>
          <menuitem label="item2" id="m_item2"/>
        </menupopup>
      </menu>

      <vbox id="debug"/>
    </vbox>
  </hbox>

</window>
